[cdc] Refactor CDC table synchronization action by extracting common parts (#2155)
